refactor: HomeCardList 분리
