+++ src/components/page/home/HomeCardList/index.tsx
@@ -1,74 +1,14 @@
 import { GroupBrowsingCardResponse } from '@api/API_LEGACY/meeting';
-import HomeCard from '@components/page/home/HomeCardList/Card';
-import { styled } from 'stitches.config';
+import CardList from '@components/page/home/HomeCardList/CardList';
 
-type HomeCardProps = {
-  label: string;
-  isMore?: boolean;
-  onMoreClick?: () => void;
-  data: GroupBrowsingCardResponse;
-};
-
-const HomeCardList = ({ label, isMore = false, onMoreClick = () => {}, data }: HomeCardProps) => {
+const HomeCardList = ({ groupBrowsingCardData }: { groupBrowsingCardData: GroupBrowsingCardResponse }) => {
   return (
-    <SCardListWrapper>
-      <STitleWrapper>
-        <STitleStyle>{label}</STitleStyle>
-        {isMore && <SMoreBtn onClick={onMoreClick}>{'더보기 >'}</SMoreBtn>}
-      </STitleWrapper>
-      <SCardWrapper>
-        {data.map(d => (
-          <HomeCard
-            key={d.id}
-            id={d.id}
-            imageURL={d.imageURL[0].url}
-            title={d.title}
-            ownerName={d.user.name}
-            ownerImage={d.user.profileImage}
-            approvedCount={d.approvedCount || 0}
-            capacity={d.capacity}
-            category={d.category}
-            canJoinOnlyActiveGeneration={d.canJoinOnlyActiveGeneration}
-            joinableParts={d.joinableParts}
-          />
-        ))}
-      </SCardWrapper>
-    </SCardListWrapper>
+    <div>
+      <CardList label="🔹 우리... 같이 솝커톤 할래?" isMore data={groupBrowsingCardData.slice(0, 3)} />
+      <CardList label="🔥 지금 모집중인 모임" data={groupBrowsingCardData.slice(0, 3)} />
+      <CardList label="🍀 1차 행사 신청이 얼마 남지 않았어요!" data={groupBrowsingCardData.slice(0, 3)} />
+    </div>
   );
 };
 
 export default HomeCardList;
-
-const SCardListWrapper = styled('section', {
-  display: 'flex',
-  flexDirection: 'column',
-
-  width: '894px',
-  paddingBottom: '$80',
-});
-
-const STitleWrapper = styled('div', {
-  display: 'flex',
-  justifyContent: 'space-between',
-  alignItems: 'center',
-
-  width: '100%',
-  paddingBottom: '$20',
-});
-
-const STitleStyle = styled('p', {
-  padding: '4px 0 8px',
-
-  fontStyle: 'H1',
-  color: '$white',
-});
-
-const SMoreBtn = styled('button', {
-  fontStyle: 'B2',
-  color: '$gray200',
-});
-
-const SCardWrapper = styled('div', {
-  display: 'flex',
-  gap: '20px',
-});
